Restructure the project in the standard way, to allow having multiple binaries
