+++ src/lib.rs
@@ -1,0 +1,4 @@
+pub mod collection;
+pub mod analysis;
+pub mod models;
+pub mod error;
